# tools/openapi.yaml
openapi: 3.0.0
info:
  title: Parking Fee Calculation API
  version: 0.1.0
servers:
  - url: /
paths:
...
  /api/calendar:
    get:
      summary: 取得日曆
      responses:
        '200':
          description: 成功
    post:
      summary: 儲存日曆
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: 成功
  /api/calendar/sync_official_v2:
    post:
      summary: 同步官方與 Nager 假日
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: 成功
  /api/calendar/generate_weekends:
    post:
      summary: 生成某年的週末清單
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: 成功